requirements.txt
youtube-transcript-api
uagents
uagents-ai-engine
# parse_qs
# urlparse
tortoise
tortoise-orm